requirements.txt
# requirements.txt
nextcord
python-dotenv
requests
beautifulsoup4
lxml